sidecar/pyproject.toml
[project]
name = "sewer-sidecar"
version = "1.0.0"
description = "Multi-Model Vision Pipeline Sidecar for Sewer-Studio"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "torch>=2.1",
    "torchvision>=0.16",
    "ultralytics>=8.2",
    "Pillow>=10.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dino = ["groundingdino-py>=0.4"]
sam = ["segment-anything-2>=0.1"]
dev = ["pytest>=8.0", "httpx>=0.27"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
